Solved some minor logical issues

+++ Learning_from_a_course/Day23-Number_system/Exercise/1_area_of_a_Circle.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 using namespace std;
 
-int area_of_a_circle(int r)
+float area_of_a_circle(int r)
 {
     // The formula to print area of a circle is pi r square
 
-    int area = 3.14 * r * r;
+    float area = 3.14 * r * r;
 
     return area;
 }
@@ -17,7 +17,7 @@ int main()
     cin >> radius;
 
     // Now call the function
-    int result = area_of_a_circle(radius);
+    float result = area_of_a_circle(radius);
     cout << "The area of a circle is : " << result << endl;
 
     return 0;
